feat: add HasPrefix method
This commit introduces the `HasPrefix` method to the `stringx` package's `String` type.

The `HasPrefix` method checks if a given `String` starts with the specified prefix. It returns `true` if the `String` has the prefix and `false` otherwise. This provides a convenient way to determine if a string begins with a particular substring.