feat: add Reverse method to String type
This commit introduces the `Reverse` method to the `String` type. This method reverses the string represented by the `String` object and returns a new `String` object containing the reversed string. It iterates over the string from the last character to the first, concatenating each character to build the reversed string.